Revise and imporve PyGraphina package structure
