+++ pygraphina/pygraphina/__init__.py
@@ -1,0 +1,5 @@
+from .pygraphina import *
+
+__doc__ = pygraphina.__doc__
+if hasattr(pygraphina, "__all__"):
+    __all__ = pygraphina.__all__
