+++ pygraphina/pygraphina/__init__.pyi
@@ -4,9 +4,22 @@ Type stubs for PyGraphina - Python bindings for Graphina graph library.
 This file provides type hints for IDEs and type checkers.
 """
 
-from typing import Optional, Dict, List, Tuple, Union, Any, Iterator
+from typing import Optional, Dict, List, Tuple, Union, Any, Iterator, final
 
+# Re-export submodules so `import pygraphina.centrality` and attribute access
+# (pygraphina.centrality) both resolve as modules for type checkers.
+from . import approximation as approximation
+from . import centrality as centrality
+from . import community as community
+from . import core as core
+from . import links as links
+from . import metrics as metrics
+from . import mst as mst
+from . import parallel as parallel
+from . import subgraphs as subgraphs
+from . import traversal as traversal
 
+@final
 class PyGraph:
     """
     A Python-accessible Graph class wrapping Graphina's core undirected graph.
@@ -385,11 +398,11 @@ class PyGraph:
         """Return the number of nodes in the graph."""
         ...
 
-    def __contains__(self, node: int) -> bool:
+    def __contains__(self, node: int, /) -> bool:
         """Check if a node exists in the graph."""
         ...
 
-
+@final
 class PyDiGraph:
     """
     A Python-accessible DiGraph class for directed graphs.
@@ -676,7 +689,7 @@ class PyDiGraph:
         """Get a view of all edges in the graph as (source, target, weight) tuples."""
         ...
 
-
+@final
 class NodeView:
     """A view for iterating over graph nodes."""
 
@@ -688,19 +701,19 @@ class NodeView:
         """Get the number of nodes."""
         ...
 
-    def __contains__(self, node: int) -> bool:
+    def __contains__(self, node: int, /) -> bool:
         """Check if a node exists."""
         ...
 
-    def __getitem__(self, node: int) -> Dict[str, int]:
+    def __getitem__(self, node: int, /) -> Dict[str, int]:
         """Get node data as a dictionary."""
         ...
 
-
+@final
 class DegreeView:
     """A view for accessing node degrees."""
 
-    def __getitem__(self, node: int) -> int:
+    def __getitem__(self, node: int, /) -> int:
         """Get the degree of a node."""
         ...
 
@@ -712,7 +725,7 @@ class DegreeView:
         """Get the number of nodes."""
         ...
 
-
+@final
 class EdgeView:
     """A view for iterating over graph edges."""
 
@@ -724,8 +737,6 @@ class EdgeView:
         """Get the number of edges."""
         ...
 
-
-# Module-level functions
 def from_networkx(nx_graph: Any) -> Union[PyGraph, PyDiGraph]:
     """
     Convert a NetworkX graph to PyGraphina.
@@ -737,7 +748,6 @@ def from_networkx(nx_graph: Any) -> Union[PyGraph, PyDiGraph]:
         A PyGraph for undirected graphs, PyDiGraph for directed graphs
     """
     ...
-
 
 def to_networkx(obj: Union[PyGraph, PyDiGraph]) -> Any:
     """
@@ -751,532 +761,117 @@ def to_networkx(obj: Union[PyGraph, PyDiGraph]) -> Any:
     """
     ...
 
-
 def to_node_dataframe(obj: Union[PyGraph, PyDiGraph]) -> Any:
     """Convert a graph's nodes to a pandas DataFrame with columns 'node_id' and 'attr'."""
     ...
-
 
 def to_edge_dataframe(obj: Union[PyGraph, PyDiGraph]) -> Any:
     """Convert a graph's edges to a pandas DataFrame with columns 'source', 'target', and 'weight'."""
     ...
 
-
-# Graph generators (also available under the core submodule)
 def complete_graph(n: int) -> PyGraph:
     """Generate a complete graph with n nodes."""
     ...
-
 
 def cycle_graph(n: int) -> PyGraph:
     """Generate a cycle graph where the n nodes form a ring."""
     ...
 
-
 def star_graph(n: int) -> PyGraph:
     """Generate a star graph with one central node connected to all others."""
     ...
-
 
 def erdos_renyi(n: int, p: float, seed: int) -> PyGraph:
     """Generate an Erdős-Rényi random graph."""
     ...
 
-
 def barabasi_albert(n: int, m: int, seed: int) -> PyGraph:
     """Generate a Barabási-Albert scale-free network."""
     ...
-
 
 def watts_strogatz(n: int, k: int, beta: float, seed: int) -> PyGraph:
     """Generate a Watts-Strogatz small-world network."""
     ...
 
-
 def bipartite(n1: int, n2: int, p: float, seed: int) -> PyGraph:
     """Generate a random bipartite graph with parts of size n1 and n2."""
     ...
 
-
-# Metrics convenience functions (also available as PyGraph/PyDiGraph methods)
 def diameter(graph: PyGraph) -> Optional[int]:
     """Compute the diameter of the graph. None if the graph is empty or disconnected."""
     ...
-
 
 def radius(graph: PyGraph) -> Optional[int]:
     """Compute the radius of the graph. None if the graph is empty or disconnected."""
     ...
 
-
 def transitivity(graph: PyGraph) -> float:
     """Compute the transitivity (global clustering coefficient) of the graph."""
     ...
-
 
 def average_clustering(graph: PyGraph) -> float:
     """Compute the average clustering coefficient of the graph."""
     ...
 
-
-# Minimum spanning tree functions
 def prim_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
     """Compute the minimum spanning tree using Prim's algorithm. Returns (total_weight, edges)."""
     ...
-
 
 def kruskal_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
     """Compute the minimum spanning tree using Kruskal's algorithm. Returns (total_weight, edges)."""
     ...
 
-
 def boruvka_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
     """Compute the minimum spanning tree using Borůvka's algorithm (parallel). Returns (total_weight, edges)."""
     ...
 
-
-# Parallel algorithm functions
 def bfs_parallel(graph: Union[PyGraph, PyDiGraph], starts: List[int]) -> List[List[int]]:
     """Run BFS from multiple starting nodes in parallel, one visit-order list per start."""
     ...
-
 
 def degrees_parallel(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, int]:
     """Compute the degree of every node in parallel."""
     ...
 
-
 def connected_components_parallel(graph: PyGraph) -> Dict[int, int]:
     """Compute a node-to-component mapping in parallel."""
     ...
 
-
-# Approximation functions
 def max_clique(py_graph: PyGraph) -> List[int]:
     """Find a large clique using a greedy heuristic."""
     ...
-
 
 def clique_removal(py_graph: PyGraph) -> List[List[int]]:
     """Partition the graph into cliques by repeated greedy clique removal."""
     ...
 
-
 def large_clique_size(py_graph: PyGraph) -> int:
     """Estimate the size of a large clique in the graph."""
     ...
-
 
 def min_weighted_vertex_cover(py_graph: PyGraph) -> List[int]:
     """Find a vertex cover using a greedy 2-approximation."""
     ...
 
-
 def average_clustering_approx(py_graph: PyGraph) -> float:
     """Estimate the average clustering coefficient by sampling."""
     ...
-
 
 def ramsey_r2(py_graph: PyGraph) -> Tuple[List[int], List[int]]:
     """Return a clique and an independent set via the Ramsey R(2, t) heuristic."""
     ...
 
-
-# Submodules
-class centrality:
-    """Centrality algorithms module."""
-
-    @staticmethod
-    def degree(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute degree centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def in_degree(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute in-degree centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def out_degree(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute out-degree centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def betweenness(graph: Union[PyGraph, PyDiGraph], normalized: bool) -> Dict[int, float]:
-        """Compute shortest-path betweenness centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def edge_betweenness(
-        graph: Union[PyGraph, PyDiGraph],
-        normalized: bool
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute betweenness centrality for all edges."""
-        ...
-
-    @staticmethod
-    def closeness(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute closeness centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def harmonic(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute harmonic centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def pagerank(
-        graph: Union[PyGraph, PyDiGraph],
-        damping: float = 0.85,
-        max_iter: int = 100,
-        tolerance: float = 1e-6,
-        nstart: Optional[Dict[int, float]] = None
-    ) -> Dict[int, float]:
-        """Compute PageRank centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def personalized_pagerank(
-        graph: Union[PyGraph, PyDiGraph],
-        personalization: Optional[List[float]] = None,
-        damping: float = 0.85,
-        tolerance: float = 1e-6,
-        max_iter: int = 100,
-        nstart: Optional[Dict[int, float]] = None
-    ) -> Dict[int, float]:
-        """Compute personalized PageRank with an optional personalization vector."""
-        ...
-
-    @staticmethod
-    def eigenvector(
-        graph: Union[PyGraph, PyDiGraph],
-        max_iter: int,
-        tolerance: float
-    ) -> Dict[int, float]:
-        """Compute eigenvector centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def katz(
-        graph: Union[PyGraph, PyDiGraph],
-        alpha: float,
-        max_iter: int,
-        tolerance: float
-    ) -> Dict[int, float]:
-        """Compute Katz centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def local_reaching_centrality(
-        graph: Union[PyGraph, PyDiGraph],
-        distance: int
-    ) -> Dict[int, float]:
-        """Compute local reaching centrality for all nodes."""
-        ...
-
-    @staticmethod
-    def global_reaching_centrality(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, float]:
-        """Compute global reaching centrality for the graph."""
-        ...
-
-
-class community:
-    """Community detection algorithms module."""
-
-    @staticmethod
-    def label_propagation(py_graph: PyGraph, max_iter: int, seed: Optional[int] = None) -> Dict[int, int]:
-        """Detect communities using the label propagation algorithm."""
-        ...
-
-    @staticmethod
-    def louvain(py_graph: PyGraph, seed: Optional[int] = None) -> List[List[int]]:
-        """Detect communities using the Louvain method."""
-        ...
-
-    @staticmethod
-    def girvan_newman(py_graph: PyGraph, target_communities: int) -> List[List[int]]:
-        """Detect communities using the Girvan-Newman algorithm."""
-        ...
-
-    @staticmethod
-    def spectral_clustering(py_graph: PyGraph, k: int, seed: Optional[int] = None) -> List[List[int]]:
-        """Partition the graph into k clusters using spectral clustering."""
-        ...
-
-    @staticmethod
-    def connected_components(py_graph: PyGraph) -> List[List[int]]:
-        """Find all connected components in the graph."""
-        ...
-
-
-class links:
-    """Link prediction algorithms module."""
-
-    @staticmethod
-    def jaccard_coefficient(
-        py_graph: PyGraph,
-        ebunch: Optional[List[Tuple[int, int]]] = None
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute the Jaccard coefficient for node pairs."""
-        ...
-
-    @staticmethod
-    def adamic_adar_index(
-        py_graph: PyGraph,
-        ebunch: Optional[List[Tuple[int, int]]] = None
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute the Adamic-Adar index for node pairs."""
-        ...
-
-    @staticmethod
-    def preferential_attachment(
-        py_graph: PyGraph,
-        ebunch: Optional[List[Tuple[int, int]]] = None
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute preferential attachment scores for node pairs."""
-        ...
-
-    @staticmethod
-    def resource_allocation_index(
-        py_graph: PyGraph,
-        ebunch: Optional[List[Tuple[int, int]]] = None
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute the resource allocation index for node pairs."""
-        ...
-
-    @staticmethod
-    def common_neighbor_centrality(
-        py_graph: PyGraph,
-        alpha: float,
-        ebunch: Optional[List[Tuple[int, int]]] = None
-    ) -> Dict[Tuple[int, int], float]:
-        """Compute common neighbor centrality for node pairs."""
-        ...
-
-    @staticmethod
-    def common_neighbors(py_graph: PyGraph, u: int, v: int) -> int:
-        """Count the number of common neighbors between two nodes."""
-        ...
-
-
-class core:
-    """Core graph generation and I/O functions."""
-
-    @staticmethod
-    def complete_graph(n: int) -> PyGraph:
-        """Generate a complete graph with n nodes."""
-        ...
-
-    @staticmethod
-    def cycle_graph(n: int) -> PyGraph:
-        """Generate a cycle graph where the n nodes form a ring."""
-        ...
-
-    @staticmethod
-    def star_graph(n: int) -> PyGraph:
-        """Generate a star graph with one central node connected to all others."""
-        ...
-
-    @staticmethod
-    def erdos_renyi(n: int, p: float, seed: int) -> PyGraph:
-        """Generate an Erdős-Rényi random graph."""
-        ...
-
-    @staticmethod
-    def barabasi_albert(n: int, m: int, seed: int) -> PyGraph:
-        """Generate a Barabási-Albert scale-free network."""
-        ...
-
-    @staticmethod
-    def watts_strogatz(n: int, k: int, beta: float, seed: int) -> PyGraph:
-        """Generate a Watts-Strogatz small-world network."""
-        ...
-
-    @staticmethod
-    def bipartite(n1: int, n2: int, p: float, seed: int) -> PyGraph:
-        """Generate a random bipartite graph with parts of size n1 and n2."""
-        ...
-
-
-class approximation:
-    """Approximation algorithms for NP-hard problems.
-
-    All functions operate on undirected graphs (PyGraph).
-    """
-
-    @staticmethod
-    def max_clique(py_graph: PyGraph) -> List[int]:
-        """Find a large clique using a greedy heuristic."""
-        ...
-
-    @staticmethod
-    def clique_removal(py_graph: PyGraph) -> List[List[int]]:
-        """Partition the graph into cliques by repeated greedy clique removal."""
-        ...
-
-    @staticmethod
-    def large_clique_size(py_graph: PyGraph) -> int:
-        """Estimate the size of a large clique in the graph."""
-        ...
-
-    @staticmethod
-    def maximum_independent_set(py_graph: PyGraph) -> List[int]:
-        """Find a maximal independent set using a greedy heuristic."""
-        ...
-
-    @staticmethod
-    def min_weighted_vertex_cover(py_graph: PyGraph) -> List[int]:
-        """Find a vertex cover using a greedy 2-approximation."""
-        ...
-
-    @staticmethod
-    def densest_subgraph(py_graph: PyGraph) -> List[int]:
-        """Find an approximately densest subgraph by greedy peeling."""
-        ...
-
-    @staticmethod
-    def average_clustering_approx(py_graph: PyGraph) -> float:
-        """Estimate the average clustering coefficient by sampling."""
-        ...
-
-    @staticmethod
-    def ramsey_r2(py_graph: PyGraph) -> Tuple[List[int], List[int]]:
-        """Return a clique and an independent set via the Ramsey R(2, t) heuristic."""
-        ...
-
-    @staticmethod
-    def local_node_connectivity(py_graph: PyGraph, source: int, target: int) -> int:
-        """Approximate the local node connectivity between source and target."""
-        ...
-
-    @staticmethod
-    def treewidth_min_degree(py_graph: PyGraph) -> Tuple[int, List[int]]:
-        """Compute a treewidth upper bound and elimination ordering using the min-degree heuristic."""
-        ...
-
-    @staticmethod
-    def treewidth_min_fill_in(py_graph: PyGraph) -> Tuple[int, List[int]]:
-        """Compute a treewidth upper bound and elimination ordering using the min-fill-in heuristic."""
-        ...
-
-
-class mst:
-    """Minimum spanning tree algorithms.
-
-    Each function operates on an undirected graph (PyGraph) and returns a tuple
-    of (total_weight, edges), where edges is a list of (u, v, weight) triples.
-    """
-
-    @staticmethod
-    def prim_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
-        """Compute the minimum spanning tree using Prim's algorithm."""
-        ...
-
-    @staticmethod
-    def kruskal_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
-        """Compute the minimum spanning tree using Kruskal's algorithm."""
-        ...
-
-    @staticmethod
-    def boruvka_mst(graph: PyGraph) -> Tuple[float, List[Tuple[int, int, float]]]:
-        """Compute the minimum spanning tree using Borůvka's algorithm (parallel)."""
-        ...
-
-
-class parallel:
-    """Parallel (Rayon-backed) graph algorithms."""
-
-    @staticmethod
-    def bfs_parallel(graph: Union[PyGraph, PyDiGraph], starts: List[int]) -> List[List[int]]:
-        """Run BFS from multiple starting nodes in parallel, one visit-order list per start."""
-        ...
-
-    @staticmethod
-    def degrees_parallel(graph: Union[PyGraph, PyDiGraph]) -> Dict[int, int]:
-        """Compute the degree of every node in parallel."""
-        ...
-
-    @staticmethod
-    def connected_components_parallel(graph: PyGraph) -> Dict[int, int]:
-        """Compute a node-to-component mapping in parallel."""
-        ...
-
-    @staticmethod
-    def pagerank_parallel(
-        graph: Union[PyGraph, PyDiGraph],
-        damping: float = 0.85,
-        max_iterations: int = 100,
-        tolerance: float = 1e-6,
-        nstart: Optional[Dict[int, float]] = None
-    ) -> Dict[int, float]:
-        """Compute PageRank scores in parallel."""
-        ...
-
-    @staticmethod
-    def triangles_parallel(graph: PyGraph) -> Dict[int, int]:
-        """Count the triangles incident to every node in parallel."""
-        ...
-
-    @staticmethod
-    def clustering_coefficients_parallel(graph: PyGraph) -> Dict[int, float]:
-        """Compute the local clustering coefficient of every node in parallel."""
-        ...
-
-    @staticmethod
-    def shortest_paths_parallel(
-        graph: Union[PyGraph, PyDiGraph],
-        sources: List[int]
-    ) -> List[Dict[int, int]]:
-        """Compute unweighted shortest path distances (hop counts) from multiple sources in parallel."""
-        ...
-
-
-class metrics:
-    """Graph metrics namespace.
-
-    The metric functions are exposed as methods on PyGraph and PyDiGraph (for
-    example, ``graph.diameter()`` and ``graph.transitivity()``). The graph-level
-    metrics ``diameter``, ``radius``, ``transitivity``, and ``average_clustering``
-    are also available as top-level functions. This module is a namespace
-    placeholder reserved for future standalone functions.
-    """
-
-
-class traversal:
-    """Graph traversal namespace.
-
-    The traversal algorithms are exposed as methods on PyGraph and PyDiGraph
-    (for example, ``graph.bfs(start)``, ``graph.dfs(start)``, ``graph.iddfs(...)``,
-    and ``graph.bidirectional_search(...)``). This module is a namespace
-    placeholder reserved for future standalone functions.
-    """
-
-
-class subgraphs:
-    """Subgraph extraction namespace.
-
-    The subgraph operations are exposed as methods on PyGraph and PyDiGraph (for
-    example, ``graph.subgraph(...)``, ``graph.ego_graph(...)``,
-    ``graph.k_hop_neighbors(...)``, and ``graph.connected_component(...)``). This
-    module is a namespace placeholder reserved for future standalone functions.
-    """
-
-
-# Type aliases (Pythonic names without the "Py" prefix)
 Graph = PyGraph
+
 DiGraph = PyDiGraph
 
-
-# Exception classes
 class GraphinaError(Exception):
     """Base exception for Graphina errors."""
     ...
 
-
 class ConvergenceError(GraphinaError):
     """Raised when an iterative algorithm fails to converge."""
     ...
-
 
 class NodeNotFoundError(GraphinaError):
     """Raised when a referenced node does not exist in the graph."""
