+++ pygraphina/pygraphina/metrics.pyi
@@ -1,0 +1,8 @@
+"""Graph metrics namespace.
+
+The metric functions are exposed as methods on PyGraph and PyDiGraph (for
+example, ``graph.diameter()`` and ``graph.transitivity()``). The graph-level
+metrics ``diameter``, ``radius``, ``transitivity``, and ``average_clustering``
+are also available as top-level functions. This module is a namespace
+placeholder reserved for future standalone functions.
+"""
