+++ pygraphina/pygraphina/subgraphs.pyi
@@ -1,0 +1,7 @@
+"""Subgraph extraction namespace.
+
+The subgraph operations are exposed as methods on PyGraph and PyDiGraph (for
+example, ``graph.subgraph(...)``, ``graph.ego_graph(...)``,
+``graph.k_hop_neighbors(...)``, and ``graph.connected_component(...)``). This
+module is a namespace placeholder reserved for future standalone functions.
+"""
