+++ pygraphina/pygraphina/traversal.pyi
@@ -1,0 +1,7 @@
+"""Graph traversal namespace.
+
+The traversal algorithms are exposed as methods on PyGraph and PyDiGraph
+(for example, ``graph.bfs(start)``, ``graph.dfs(start)``, ``graph.iddfs(...)``,
+and ``graph.bidirectional_search(...)``). This module is a namespace
+placeholder reserved for future standalone functions.
+"""
