+++ pygraphina/pyproject.toml
@@ -44,8 +44,8 @@ bindings = "pyo3"
 features = ["pyo3/extension-module"]
 include = [
     { path = ".", format = "sdist" },
-    { path = "pygraphina.pyi", format = ["sdist", "wheel"] },
-    { path = "py.typed", format = ["sdist", "wheel"] }
+    { path = "pygraphina/*.pyi", format = ["sdist", "wheel"] },
+    { path = "pygraphina/py.typed", format = ["sdist", "wheel"] }
 ]
 strip = true
 release = true
